+++ src/app/_components/home/home.component.ts
@@ -1,9 +1,11 @@
-import { Component } from '@angular/core';
+import { Component, HostListener, inject } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { AgentsComponent } from '../agents/agents.component';
 import { QueuesComponent } from '../queues/queues.component';
 import { PanelStatusEnum } from '../../_models';
 import { DrawerModule } from 'primeng/drawer';
+import { FilterMetadata } from 'primeng/api';
+import { NotificationApiService } from '../../_services';
 
 @Component({
   selector: 'app-home',
@@ -21,6 +23,22 @@ export class HomeComponent {
   agentStatus: PanelStatusEnum = PanelStatusEnum.EXPANDED;
   isVisible = false;
   protected readonly panelStatus = PanelStatusEnum;
+  filters: {
+    [s: string]: FilterMetadata | FilterMetadata[];
+  } = {};
+  sortOrder: number = -1;
+  private readonly notificationService = inject(NotificationApiService);
+
+  @HostListener('document:visibilitychange')
+  visibilitychange() {
+    if (!document.hidden) {
+      const state = this.notificationService.getSocketState();
+      if (state !== 1) {
+        console.log(`[Notifications][HealthCheck] Initializing WS`);
+        this.notificationService.notificationInitialize();
+      }
+    }
+  }
 
   toggleSidebar() {
     this.isVisible = !this.isVisible;
@@ -34,4 +52,5 @@ export class HomeComponent {
       this.queueStatus = PanelStatusEnum.COLLAPSED;
     }
   }
+
 }