+++ src/app/_services/api/conversations-api.service.ts
@@ -11,4 +11,32 @@ export class ConversationsApiService {
     return from(this.apiInstance.postConversationsCallParticipantMonitor(conversationId, participantId));
   }
 
+  queueConversationsQuery(queueIds: string[]) {
+
+    const opts = {
+      order: 'asc',
+      filter: {
+        type: 'or',
+        predicates: queueIds.map(id => ({
+          type: 'dimension',
+          dimension: 'queueId',
+          value: id
+        }))
+      },
+      groupBy:['queueId'],
+      metrics: [
+        { metric: 'oWaiting', details: true }
+      ],
+      subscribe: true
+    };
+
+    return from(this.apiInstance.postAnalyticsConversationsActivityQuery(opts));
+  }
+
+  assignConversation(conversationId: string, userId: string) {
+    return from(this.apiInstance.postConversationAssign(conversationId, {
+      id: userId
+    }));
+  }
+
 }